feat #158: 수정중 임시로 커밋함

+++ walloud_back/src/main/java/com/spring/mydiv/Service/UserService.java
@@ -5,6 +5,7 @@ import javax.transaction.Transactional;
 import com.spring.mydiv.Dto.*;
 import com.spring.mydiv.Entity.Person;
 import com.spring.mydiv.Exception.DefaultException;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.spring.mydiv.Entity.User;
@@ -19,6 +20,7 @@ import java.util.List;
 import java.util.Optional;
 
 import static com.spring.mydiv.Code.ErrorCode.*;
+import static java.lang.Boolean.TRUE;
 
 /**
  * @author 12nov
@@ -30,28 +32,34 @@ public class UserService {
 	private final PersonRepository personRepository;
 	private final TravelRepository travelRepository;
     private final S3UploaderService s3UploaderService;
-	
+    private final TravelService travelservice;
+    private final PersonService personservice;
+
     @Transactional
     public UserDto.Response createUser(UserDto.Request request) {
-        User user = User.builder()
-                .name(request.getUser_name())
-                .email(request.getUser_email())
-                .password(request.getUser_password())
-                .account(request.getUser_account())
-                .bank(request.getUser_bank())
-                .build();
-        userRepository.save(user);
-        return UserDto.Response.fromEntity(user);
-    } //fin
+        if (!checkIsEmailRegistered(request.getUser_email())){
+            User user = User.builder()
+                    .name(request.getUser_name())
+                    .email(request.getUser_email())
+                    .password(request.getUser_password())
+                    .account(request.getUser_account())
+                    .bank(request.getUser_bank())
+                    .build();
+            userRepository.save(user);
+            return UserDto.Response.fromEntity(user);
+        }
+        else
+            throw new DefaultException(ALREADY_REGISTERED);
+    }
 
     public boolean checkIsEmailRegistered(String email){
         return userRepository.existsByEmail(email);
     }
-    public UserDto.ResponseWithImage login(UserDto.Login loginUser) {
+    public UserDto.Response login(UserDto.Login loginUser) {
         User entity = userRepository.findByEmail(loginUser.getEmail())
                 .orElseThrow(() -> new DefaultException(WRONG_EMAIL));
         if (loginUser.getPassword().equals(entity.getPassword()))
-            return UserDto.ResponseWithImage.fromEntity(entity);
+            return UserDto.Response.fromEntity(entity);
         else throw new DefaultException(WRONG_PASSWORD);
     } //ing
 
@@ -73,7 +81,7 @@ public class UserService {
             result.add(travel);
         }
         return result;
-    } //fin
+    }
 
 
 
@@ -107,6 +115,7 @@ public class UserService {
 
         if (updateRequest.getUser_name() != null) user.setName(updateRequest.getUser_name());
         if (updateRequest.getUser_email() != null) user.setEmail(updateRequest.getUser_email());
+        // TODO 이메일 중복인지 확인해야할 것 같음
         if (updateRequest.getUser_password() != null) user.setPassword(updateRequest.getUser_password());
         if (updateRequest.getUser_account() != null) user.setAccount(updateRequest.getUser_account());
         if (updateRequest.getUser_bank() != null) user.setBank(updateRequest.getUser_bank());
@@ -118,9 +127,13 @@ public class UserService {
     public UserDto.ResponseWithImage updateUserImage(int userId, String imageURL){
         User user = userRepository.findById(Long.valueOf(userId))
                 .orElseThrow(() -> new DefaultException(NO_USER));
-//        deleteUserImage(user);
+        // TODO - deleteUserImage(user);
         user.setInfo(imageURL);
         return UserDto.ResponseWithImage.fromEntity(userRepository.save(user));
+        userRepository.findById(Long.valueOf(userId))
+                .map(UserDto.ResponseWithImage::fromEntity)
+                .orElseThrow(()-> new DefaultException(NO_USER))
+                .getImageurl();
     }
 
     public void deleteUserImage(User user){
@@ -136,9 +149,22 @@ public class UserService {
     }
 
     public void deleteUser(int userId){
-        userRepository.deleteById(Long.valueOf(userId));
+        if(getUserJoinedTravel(userId).size() == 0)
+            // can delete this user - not participated some travel
+            userRepository.deleteById(Long.valueOf(userId));
+        else
+            throw new DefaultException(INVALID_DELETE_TRAVEL_EXISTED);
     }
 
+    public int createNewTravelUserJoining(int userId, String travelName) {
+        TravelDto.Request travelRequest = new TravelDto.Request(travelName);
+        PersonDto.Request personRequest = new PersonDto.Request(
+                getUserInfo(userId),
+                travelservice.createTravel(travelRequest));
+        if (ResponseEntity.ok(personservice.createPerson(personRequest, TRUE)).getStatusCodeValue() == 200)
+            return personRequest.getTravel().getTravelId().intValue();
+        else throw new DefaultException(CREATE_FAIL);
+    }
 }
 
 //
